Implement upload file on the server for #12.
The editor router of the server gets a new function called 'uploadFile'.
It expect the content send as form data with a path and a file.
Use the functionality of the 'koa-body' module, that can store form data
files automatically.
Move the file from the temporally storage to the clients source
directory with the defined relative path.