Make advertisement banner moveable #5.
The image of the advertisement can be shifted vertically and horizontal.
Shifting is possible by using the related buttons or dragging.
Add package 'v-dragged' to the dependencies to handle the dragging.
The possibility of dragging is shown by different cursor styles.
On desktop, the image will be stretched to the height and can be shifted
on the horizontal axis.
On mobile, the image will be stretched to the width and can be shifted
to the vertical axis.
It is not posisble to shift the image over any of the borders. A red
blink is shown to the user as a warning, that this motion is not
possible.
Some cosmetics to the button group of upload an image and shift it.