Introduce the vue-router to switch subpages.
Complete rewrite this procedure.
Add the vue-router module to the dependencies and add it to the Vue
instance.
Introduce the Router class, which configures the vue-router.
Export the background definition from the main menu data set to the
router configuration. By this the app can directly access the background
for the current route and no connection to the main menu is necessary
anymore.
Introduce a new enumerations RouteNames to use named routes, which are
used by the router and the main menu. For this the DataManager has been
extended to replace such entries.