Introduce the EditWindow component.
Extend the subpage mixin to provide the editText() function.
Calling this function with a reference to a data object, it display the
current value and let the user change it.
Changes are applied like a model property.
The operation can be aborted (changes reset) or finished.
Export Bulmas modal card style adjustments to the 3v-forhair style
sheet.
Add new labels for the EditWindow component.
Small adjustments to subpages for testing this new functionality.