Make the whole update process working.
Extend the api connector by the update config function.
Fix issues to handle backend responses.
Fix issue in the backend to retrieve the session key correctly.
Add live view of the build process to the backend.
Add global event to start save process for all components.
Mixin a default save function for each subpage to update its data object
in the backend and call the final build afterwards.
Provide a final build function which call the backend to rebuild the
client.
Fix false entry to the data store entries.