Add DataProtectionHint component.
Export the box to show the data protection hint to its own component,
cause it will be used in several places.
It consists of a button, which if clicked opens the info box.
Extend the labels and their keys for that component.
Add a new data object for the content of the box and add a new data
store entry for that.
Inject the component into the Impressum page so far.