Fix the position of the header line.
Due to the new header text and image, the positions was no longer
correct.
Due to the fact that the image is not relative, but absolutely placed in
the layout, a correct calculation seems not to be possible.
This was indicated by the comment in the style section also.
By adjusting this mysterious value, by visual checks it was possible to
make it look well again.