Add mail router to server and improve server.
Add new router to server for sending mails as visitor feedback.
Establish the mailer utilities, which handle the actually mail sending.
Add nodemailer to the dependencies.
Add new properties file for the mailer, including to read in the
password (will be ignored in repository)
Extend the core by an object which describes the request parameter keys.
Add a constant file for the routes of the server and adjust all routers.
The session check and admin view middleware are using the router names now as well.