Extend webpack's development server for test purposes.
When using the client by the development server trough webpack and
running the server concurrently to use the API for editing, this cause
unwanted reloads of the client, cause webpack recognized changes in
background.
Therefore the hot reload middleware can be disabled with the new
argument '--no-hotload'.
A new NPM script target called 'dev-static' sets this argument already.