Split client save and building for admins.
When using the editable administration view, a new button called
'build' has been established.
The already existing button 'save' now only update the data of the
subpages with the backend.
This allows the user to save frequently and interrupt work without have
to rebuild the page.
When everything is done, the build button can be used to force the
server to rebuild, which is still a long process.
The function has been renamed from 'save' to 'build' on the ApiConnector
to differ both steps.
The label keys and translations have been extended for the new button.
The subpage defaults save function due to the mixin provides a check if
any data has been adjusted and only updates if so.
This option seems to be buggy right now and does not work always for
sure.