Hardly improve session handling of the client.
The ApiConnector has been extended to store the session key to the
local storage after a succesful login.
Furthermore the ApiConnector has get an extra function called
'validateSession'.
On startup/loading the client, it checks for a cached session key.
The App component will cause a session validation on creation.
The validation will cause an login event, if no valid session key is
available.
This also clears the local cache.
This event is catched by the App, which will open the authentication
modal.
Each function that requires the session key use the session validation
to make sure it will work out.
Unfortunately if not so, a login will be required by the user, but the
action itself gets canceled and must be invoked again by the user.
The logout function delete the cached session key as well.
A verification of the session by the server is still missing and will be
implemented soon.