Two minor bug fixes.
The stylist category field was a router link, so editing the text was
not possible.
Removed the link and use function that can differ between being in
normal or editable mode.
Add a class to have a hover pointer.
Fix the issue that reloading in administration view and reusing an old
session key does not enable the editable view.