Big shape, font-size and alignment update.
Use the root font size for all relations now.
The font-size depends on the available screen size and on the device
type, but make sure to keep in a range.
Fix the footer space on mobile devices.
Make the mobile menu layover clickable to close.
Add some margins to look better.
Adjust some text alignments for the mobile view.
Increase the size of the products info button to be better clickable.
The stylist search field is full width on mobile devices now.
Header size is relative to the device now as well.