Bunch of bug fixes.
The costs data object was not correctly parseable.
By this, the procedure ignores the rest of the data object and the
images could not been loaded.
So the action images for him are available now again.
Add action carousel for her also and therefore add an image.
Restyle the action carousels by using background images instead of plain
img directives, cause their shape is better to automatable.
Adjust the main menu title positioning (shift more to left by the factor).
Add default texts for the missing product description data fields.