hotfix: advertisement banner on mobile views
The class to hide the right column of the banner on mobile devices was
not "strong" enough and was overruled by refactoring issues.

The CSS rules by Bulma for mobile devices does not take affect on large
mobile screens of modern phones. So this must be done manually.
Else the right column of the banner disappears, but the left column is
still only taking half of the screen.