remove multilanguage support and translations
First of all the origin cause for this change is that the used
`vue-multilang-store` dependency is causing issues with Uglify and ES6
standards.
But the multi-language feature has never been used in practice. While
there are several translations for static headers and titles, the
component to switch the language has been commented out. So the user was
never able to see something else than the German translations.
Furthermore is all the editable conent that is stored in the data JSON
files exclusively in German too. So it doesn't really make sense to have
multi-language support at this point in time.
Anyway, if it becomes necessary at some point in time it can be easy
added again with a proper Vue plugin like `vue-i18n`.