remove Vue plugin to detect devices
This plugin was used to alternate the behavior of the main menu
component on different devices, primarily desktop or mobile. But this
plugin was causing issues with Uglify and ES6 standards. As it was only
used very subtle and it was simple to replace it with a custom computed
property, it got removed to solve this issue.
The only resulting disadvantage is that the used number to identify the
width of a desktop device is no more strongly related to the `include
media` library that gets used for the styling sheets.